allan,

Well, I gotta say these changes look pretty good, but I'm crossing my fingers that they don't mess up the server.  So, even though I didn't review them all in detail, they mostly make sense to me and I'm merging them in.  \o/

I reviewed in detail the ones that looked questionable to me and saw they were good after all.

Nice job!

Now, let's get together in IRC and hopefully talk in Teamspeak and I think you're the next candidate to gain access to review and approve pull requests along side myself.  We just gotta lay down some more ground rules and such.

Thanks,

-- Aknor

+++ src/eve-server/EVEServerConfig.h
@@ -58,7 +58,7 @@ public:
     struct
     {
         /// Role to assign to auto created account; set to 0 to disable auto account creation.
-        uint32 autoAccountRole;
+        uint64 autoAccountRole;
         /// A message shown to every client on login.
         std::string loginMessage;
     } account;
